refactor: move ToggleDark.svelte to src/routes
The ToggleDark.svelte component was moved from the
src/routes/[_type=type]/[id] directory to the src/routes directory.
The import statement in +page.svelte was updated accordingly.
